refactor: reorganize services under FixedSessions namespace
- FixedSessions::Creator — shared session creation, no device-specific logic
- FixedSessions::AirBeamMini2::BinaryParser — binary protocol parsing (Mini2-specific)
- FixedSessions::AirBeamMini2::Ingester — binary measurement ingestion (Mini2-specific)

The FixedSessions top-level namespace will grow to serve all AirBeam models
using the new v3 endpoint. The AirBeamMini2 sub-namespace scopes the binary
protocol services without polluting the shared layer.

Use ::FixedSessions in controllers to avoid constant lookup collision with
the Api::V3::FixedSessions controller namespace.
