+++ app/services/fixed_sessions/air_beam_mini2/ingester.rb
@@ -1,5 +1,5 @@
-module AirBeamMini2
-  module Measurements
+module FixedSessions
+  module AirBeamMini2
     class Ingester
       def initialize(
         parser: BinaryParser.new,
